+++ R/tidy_name.R
@@ -1,13 +1,16 @@
 tidy_name <- function(x, split_on = "_") {
     new_spec_names <- unlist(lapply(x, function(y) {
         split_str <- unlist(stringr::str_split(y, split_on))
-            return(paste0(
-                stringr::str_to_upper(stringr::str_sub(split_str[1], 1, 1)),
-                stringr::str_sub(split_str[1], 2, nchar(split_str[1])),
-                stringr::str_to_upper(stringr::str_sub(split_str[2],1,1)),
-                stringr::str_sub(split_str[2], 2, nchar(split_str[2])),
-                collapse = ""
-            ))
+
+        first <- split_str[1]
+        second <- ifelse(is.na(split_str[2]), "", split_str[2])
+        return(paste0(
+            stringr::str_to_upper(stringr::str_sub(first, 1, 1)),
+            stringr::str_sub(first, 2, nchar(first)),
+            stringr::str_to_upper(stringr::str_sub(second, 1, 1)),
+            stringr::str_sub(second, 2, nchar(second)),
+            collapse = ""
+        ))
     }))
 
     return(new_spec_names)